Move AnswerButton to a different file
